# tests/STLCompliance.OpenApi.Tests/snapshots/routarr.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RoutArr.Api | v1"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/integrations/supplyarr-demand-status": {
      "post": {
        "tags": [
          "Integrations"
        ],
        "operationId": "IngestSupplyarrDemandStatus",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestSupplyarrDemandStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/integrations/supplyarr-demand-status": {
      "post": {
        "tags": [
          "Integrations"
        ],
        "operationId": "IngestSupplyarrDemandStatusV1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestSupplyarrDemandStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IngestSupplyarrDemandStatusRequest": {
        "required": [
          "tenantId",
          "routarrPublicationId",
          "supplyarrDemandRefId",
          "supplyarrCallbackPublicationId",
          "eventType",
          "procurementStatus",
          "supplyarrPurchaseRequestId",
          "supplyarrPurchaseOrderId",
          "supplyarrReceivingReceiptId",
          "quantityReceivedDelta",
          "message",
          "occurredAt"
        ],
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "routarrPublicationId": {
            "type": "string",
            "format": "uuid"
          },
          "supplyarrDemandRefId": {
            "type": "string",
            "format": "uuid"
          },
          "supplyarrCallbackPublicationId": {
            "type": "string",
            "format": "uuid"
          },
          "eventType": {
            "type": "string"
          },
          "procurementStatus": {
            "type": "string"
          },
          "supplyarrPurchaseRequestId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "supplyarrPurchaseOrderId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "supplyarrReceivingReceiptId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "quantityReceivedDelta": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}